+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Generics integrated into `EventDataRunner`, `NewMatchupRunner`, `MatchupOutput`, and `EventDataOutput`
+- Replacement interface{} --> any
+- New interaction of signatures for `NewEventDataRunner`and `NewMatchupRunner` to support generics pattern
 
 ## [0.19.0] - 2026-02-08
 ### Fixed